"8:55am. Enough wasting time. Let me just go through Pharo By Example.
There is Python, but the RL environments won't be hard to link up in any language that I want. The same goes for making a ML library.

The IDE experience Pharo offers on the other hand is something no other language can emulate. I'd literally need to make my own VM for that.

Even though I've decided to pick a dynamic language, Python is really a language that feels like it was designed by somebody who does not understand abstraction. Smalltalk on the other hand definitely understands it and the same goes for Pharo.

I can do RL in anything, but I cannot help it if the fucking GUI hangs the entire interpreter. Literally all the plotting examples that I've seen so far have been broken in some way.

Honestly, if the state of RL was good enough that I could just pick up Python and begin the assault, somebody else would have done this long ago. For the sake of making money the reality is that getting a job would be a lot easier than trying to go down this route. So since things are like that I should just carry on like I want and program in what I want to.

Yeah, the reality is definitely that I won't be able to cause the Singularity without the ability to make a generic Cuda map kernel. But another reality is also that I also won't be able to cause it from the command line. Spiral is the pinnacle of abstraction that a systems language can reach, but now it is time to break through to the heavens.

9:15am. http://forum.world.st/Multi-core-Software-Concurrency-td4993926.html

I am looking up stuff on Pharo and the concurrency status of it. I guess having multiple VMs communicating with each other might be doable even if it is not exactly agent based concurrency. I actually do not understand how sockets works, but this is probably something I'd need to make Spiral a persistent interpreter.

Also I am happy to see that there is something on GPU programming here.

...I won't be ashamed of what I am doing. Since I have the freedom to do so, I should just program in whatever way I really want. That is how I will win in the end. If they really can do it using Python scripts, then I'll be happy to see them do it. I am not at all sarcastic about that. I know that I have my limitations. I want to program in a way that feels good to me in the end.

Should I really be posting this in the Python Projects file? I'd rather not go there and bash it.

9:20am. Very well, I'll do it just for today and maybe tomorrow and then I'll go back to the Spiral language project. Maybe later I'll open a repo specific to Pharo."